[coupon] css : 메인화면 전반 css 수정

+++ src/main/java/hello/fclover/controller/MemberController.java
@@ -448,7 +448,7 @@ public class MemberController {
     }
 
     @GetMapping("/memberPay")
-    public String sellerPay(@RequestParam String goodsName,@RequestParam int goodsPrice,@RequestParam String goodsWriter,@RequestParam int quantity, @RequestParam int goodsNo, Principal principal, Model model) {
+    public String sellerPay(@RequestParam String goodsName,@RequestParam int goodsPrice,@RequestParam String goodsWriter,@RequestParam int quantity, @RequestParam Long goodsNo, Principal principal, Model model) {
         if (principal == null) {
             return "redirect:/login";
         }
@@ -462,6 +462,8 @@ public class MemberController {
 
         return "user/userPayments";
     }
+
+
 
     @GetMapping("/memberPayDone")
     public String sellerPayDone() {
